+++ crates/db-storage/src/lib.rs
@@ -25,6 +25,7 @@ pub mod migrations;
 pub mod module_resources;
 pub mod rooms;
 pub mod sip_configs;
+pub mod streaming_targets;
 pub mod tariffs;
 pub mod tenants;
 pub mod users;
@@ -40,5 +41,6 @@ pub mod sql_types {
     pub use super::events::EventInviteStatusType as EventInviteStatus;
     pub use diesel::sql_types::*;
     pub use types::core::InviteRoleType as InviteRole;
+    pub use types::core::StreamingKindType as StreamingKind;
     pub use types::core::TariffStatusType as TariffStatus;
 }